requirements.txt
easyAI
numpy~=2.3.1
scikit-fuzzy~=0.5.0
matplotlib
networkx
scipy
python-dotenv~=1.2.1
requests~=2.32.5
pandas~=2.3.3
scikit-learn~=1.7.2